Changed Inches, int to float

+++ TC_assignment2.c
@@ -3,7 +3,7 @@
 
 struct Distance {
     int feet;
-    int inches;
+    float inches;
 };
 
 int addFeet(int d1ft, int d2ft){
@@ -12,8 +12,8 @@ int addFeet(int d1ft, int d2ft){
     return totalft;
 }
 
-int addInches(int d1in, int d2in){
-    int totalin = 0;
+float addInches(float d1in, float d2in){
+    float totalin = 0;
     totalin = d1in + d2in;
     return totalin;
 }
@@ -28,27 +28,27 @@ printf("1st distance\n");
 printf("Input feet: ");
 scanf("%d", &d1.feet);
 printf("Input inches: ");
-scanf("%d", &d1.inches);
+scanf("%f", &d1.inches);
 
 // enter second distance input
 printf("\n2nd distance\n");
 printf("Input feet: ");
 scanf("%d", &d2.feet);
 printf("Input inches: ");
-scanf("%d", &d2.inches);
+scanf("%f", &d2.inches);
 
 // add distances
 int resultfeet;
-int resultinches;
+float resultinches;
 
 resultfeet = addFeet(d1.feet, d2.feet);
 resultinches = addInches(d1.inches, d2.inches);
 
 // convert inches to feet if greater than or equal to 12
-while (resultinches >= 12) {
-resultinches = resultinches - 12;
+while (resultinches >= 12.0) {
+resultinches = resultinches - 12.0;
 ++resultfeet;
 }
-printf("\nSum of distances = %d\'-%d\"", resultfeet, resultinches);
+printf("\nSum of distances = %d\'-%.1f\"", resultfeet, resultinches);
 return 0;
 }
